- Fixed a compiler error on the mac. - Upped the renovation cost a bit. It is still cheaper as before when doing it immediately, but raises till higher values as originally when postponing it.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@4973 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Scripting/OOJSPlayer.m
@@ -455,7 +455,7 @@ static JSBool PlayerReplaceShip(JSContext *context, uintN argc, jsval *vp)
 	if (argc > 1)
 	{
 		JS_ValueToInt32(context,OOJS_ARGV[1],&personality);
-		if (personality >= 0 && personality < ENTITY_PERSONALITY_MAX)
+		if (personality >= 0 && (uint16_t)personality < ENTITY_PERSONALITY_MAX)
 		{
 			[player setEntityPersonalityInt:(uint16_t)personality];
 		}
